mini_pupper_control_v2/scripts/requirements.txt
rich==12.5.1
pyyaml==6.0
pydantic==1.9.2
pytest==7.1.2
pytest-mock==3.8.2
